non-public plot methods moved to _plot.py
